+++ osctiny/tests/test_packages.py
@@ -1,8 +1,11 @@
+from io import StringIO, BytesIO, IOBase
+import re
 from unittest import skip
 
 import responses
 
 from .base import OscTest, CallbackFactory
+from osctiny.errors import OscError
 
 
 class TestPackage(OscTest):
@@ -99,11 +102,9 @@ class TestPackage(OscTest):
 
     @responses.activate
     def test_get_meta(self):
-        self.mock_request(
-            method=responses.GET,
-            url=self.osc.url + '/source/SUSE:SLE-12-SP1:Update/python.8549/'
-                               '_meta',
-            body="""
+        def callback(headers, params, request):
+            status = 200
+            body = """
             <package name="python.8549" project="SUSE:SLE-12-SP1:Update">
               <title>Python Interpreter</title>
               <description>Python is an interpreted, object-oriented programming
@@ -117,16 +118,45 @@ class TestPackage(OscTest):
               <releasename>python</releasename>
             </package>
             """
+            if params.get("view", None) == "blame":
+                body = """
+   1 (foo__bar     2018-10-29 16:28:55     1) <package name="python.8549" project="SUSE:SLE-12-SP1:Update">
+   1 (foo__bar     2018-10-29 16:28:55     2)   <title>Python Interpreter</title>
+   1 (foo__bar     2018-10-29 16:28:55     3)   <description>Python is an interpreted, object-oriented programming language, and is
+   1 (foo__bar     2018-10-29 16:28:55     4) often compared to Tcl, Perl, Scheme, or Java.  You can find an overview
+   1 (foo__bar     2018-10-29 16:28:55     5) of Python in the documentation and tutorials included in the python-doc
+   1 (foo__bar     2018-10-29 16:28:55     6) (HTML) or python-doc-pdf (PDF) packages.
+   1 (foo__bar     2018-10-29 16:28:55     7) 
+   1 (foo__bar     2018-10-29 16:28:55     8) If you want to install third party modules using distutils, you need to
+   1 (foo__bar     2018-10-29 16:28:55     9) install python-devel package.</description>
+   1 (foo__bar     2018-10-29 16:28:55    10)   <releasename>python</releasename>
+   1 (foo__bar     2018-10-29 16:28:55    11) </package>
+                """
+
+            return status, headers, body
+
+        self.mock_request(
+            method=responses.GET,
+            url=self.osc.url + '/source/SUSE:SLE-12-SP1:Update/python.8549/'
+                               '_meta',
+            callback=CallbackFactory(callback)
         )
 
-        response = self.osc.packages.get_meta(
-            "SUSE:SLE-12-SP1:Update", "python.8549"
-        )
-        self.assertEqual(response.tag, "package")
-        self.assertEqual(
-            response.xpath("./title")[0].text,
-            "Python Interpreter"
-        )
+        with self.subTest("without blame"):
+            response = self.osc.packages.get_meta(
+                "SUSE:SLE-12-SP1:Update", "python.8549"
+            )
+            self.assertEqual(response.tag, "package")
+            self.assertEqual(
+                response.xpath("./title")[0].text,
+                "Python Interpreter"
+            )
+
+        with self.subTest("with blame"):
+            response = self.osc.packages.get_meta(
+                "SUSE:SLE-12-SP1:Update", "python.8549", blame=True
+            )
+            self.assertTrue(isinstance(response, str))
 
     @skip("No test data available")
     @responses.activate
@@ -219,3 +249,223 @@ class TestPackage(OscTest):
         self.assertIn(
             "++#if FFI_TYPE_LONGDOUBLE != FFI_TYPE_DOUBLE", response
         )
+
+    @responses.activate
+    def test_set_meta(self):
+        bodies = []
+
+        def callback(headers, params, request):
+            bodies.append(request.body)
+            status, body = 200, ""
+
+            return status, headers, body
+
+        self.mock_request(
+            method=responses.PUT,
+            url=re.compile(self.osc.url + '/source/(?P<project>[^/]+)/'
+                                          '(?P<package>[^/]+)/_meta'),
+            callback=CallbackFactory(callback)
+        )
+
+        data = (
+            (
+                {
+                    'project': "test:project",
+                    'package': "test.package",
+                    'title': "test",
+                    'description': "foo"
+                },
+                b'<package><title>test</title><description>foo</description>'
+                b'</package>'
+            ),
+            (
+                {
+                    'project': "test:project",
+                    'package': "test.package",
+                },
+                b'<package><title/><description/></package>'
+            ),
+            (
+                {
+                    'project': "test:project",
+                    'package': "test.package",
+                    'meta': """
+                    <package name="test.package" project="test:project">
+                      <title/>
+                      <description/>
+                      <build>
+                        <enable repository="openSUSE_Leap_15.0"/>
+                        <disable arch="i586"/>
+                      </build>
+                    </package>
+                    """
+                },
+                b'<package name="test.package" project="test:project"><title/>'
+                b'<description/><build>'
+                b'<enable repository="openSUSE_Leap_15.0"/>'
+                b'<disable arch="i586"/></build></package>'
+            ),
+            (
+                {
+                    'project': "test:project",
+                    'package': "test.package",
+                    'title': 'foo',
+                    'description': 'bar',
+                    'meta': """
+                    <package name="test.package" project="test:project">
+                      <title/>
+                      <description/>
+                      <build>
+                        <enable repository="openSUSE_Leap_15.0"/>
+                        <disable arch="i586"/>
+                      </build>
+                    </package>
+                    """
+                },
+                b'<package name="test.package" project="test:project">'
+                b'<title>foo</title><description>bar</description><build>'
+                b'<enable repository="openSUSE_Leap_15.0"/>'
+                b'<disable arch="i586"/></build></package>'
+            ),
+        )
+
+        for params, expected in data:
+            with self.subTest():
+                self.osc.packages.set_meta(**params)
+                self.assertEqual(bodies[-1], expected)
+
+    @responses.activate
+    def test_push_file(self):
+        content = """
+        ლ(ಠ益ಠ)ლ            ლ(ಠ益ಠ)ლ
+        Lorem ipsum dolor sit amet,
+        consectetur adipiscing elit.
+        Vestibulum id enim 
+        fermentum, lobortis urna
+        quis, convallis justo. 
+        ლ(ಠ益ಠ)ლ            ლ(ಠ益ಠ)ლ 
+        """
+        bodies = []
+
+        def callback(headers, params, request):
+            if isinstance(request.body, IOBase):
+                request.body.seek(0)
+                bodies.append(request.body.read())
+            else:
+                bodies.append(request.body)
+            status, body = 200, ""
+
+            return status, headers, body
+
+        self.mock_request(
+            method=responses.PUT,
+            url=re.compile(
+                self.osc.url + '/source/(?P<project>[^/]+)/'
+                               '(?P<package>[^/]+)/(?P<filename>.+)'
+            ),
+            callback=CallbackFactory(callback)
+        )
+
+        with self.subTest("as unicode"):
+            self.osc.packages.push_file("prj", "pkg", "readme.txt", content)
+            self.assertEqual(bodies[-1], content.encode())
+
+        with self.subTest("as bytes"):
+            self.osc.packages.push_file("prj", "pkg", "readme.txt",
+                                        content.encode())
+            self.assertEqual(bodies[-1], content.encode())
+
+        with self.subTest("as StringIO"):
+            self.osc.packages.push_file("prj", "pkg", "readme.txt",
+                                        StringIO(content))
+            self.assertEqual(bodies[-1], content.encode())
+
+        with self.subTest("as BytesIO"):
+            self.osc.packages.push_file("prj", "pkg", "readme.txt",
+                                        BytesIO(content.encode()))
+            self.assertEqual(bodies[-1], content.encode())
+
+    @responses.activate
+    def test_aggregate(self):
+        put_called = []
+
+        def exists_callback(headers, params, request):
+            status, body = 404, ""
+            if "exists" in request.url:
+                if "_aggregate" in request.url and "already.agged" not in request.url:
+                    status = 404
+                else:
+                    status = 200
+            return status, headers, body
+
+        def put_callback(headers, params, request):
+            put_called.append({'request': request, 'params': params})
+            status, body = 200, ""
+            return status, headers, body
+
+        def meta_callback(headers, params, request):
+            status = 200
+            body = """<package><title/><description/></package>"""
+            return status, headers, body
+
+        self.mock_request(
+            method=responses.HEAD,
+            url=re.compile(
+                self.osc.url + '/source/.*'
+            ),
+            callback=CallbackFactory(exists_callback)
+        )
+        self.mock_request(
+            method=responses.PUT,
+            url=re.compile(
+                self.osc.url + '/source/.*'
+            ),
+            callback=CallbackFactory(put_callback)
+        )
+        self.mock_request(
+            method=responses.GET,
+            url=re.compile(
+                self.osc.url + '/source/.+/_meta'
+            ),
+            callback=CallbackFactory(meta_callback)
+        )
+
+        with self.subTest("identical package"):
+            self.assertRaises(
+                OscError,
+                self.osc.packages.aggregate,
+                "test:project", "test.package",
+                "test:project", "test.package",
+            )
+
+        with self.subTest("non-existing package"):
+            self.assertRaises(
+                OscError,
+                self.osc.packages.aggregate,
+                "test:project", "test.package",
+                "test:project:2", "test.package",
+            )
+
+        with self.subTest("already existing aggregate"):
+            self.assertRaises(
+                OscError,
+                self.osc.packages.aggregate,
+                "test:project:exists", "test.package",
+                "test:project2:exists", "already.agged",
+            )
+
+        with self.subTest("non-existing target package"):
+            old_len = len(put_called)
+            self.osc.packages.aggregate(
+                "test:project:exists", "test.package",
+                "test:project2:foo", "test.pkg",
+            )
+            self.assertEqual(len(put_called), old_len + 2)
+
+        with self.subTest("existing target package"):
+            old_len = len(put_called)
+            self.osc.packages.aggregate(
+                "test:project:exists", "test.package",
+                "test:project2:exists", "test.pkg",
+            )
+            self.assertEqual(len(put_called), old_len + 1)